Can now go through sign up to log in, can now log out
Sign up gets the user from the data from Shayan's backend and exposes that as currentUser
to the rest of the app. The token gets put in local storage. On start up, if a token exists
in local storage, we sign the user in. To log out, currentUser is set to null and the local
storage with the token is deleted.

Log in doesn't currently work, but can probably be implemented by calling some log in apiClient
and then setting the local storage token