Refactored and cleaned up the codebase.
Refactored explore page to handle creating StudyGroupCard join/leave button.
Created a UserGroupContext to be able to access the all study groups and the groups
the user is in in every screen in the App.
Generally cleaned up the code by getting rid of unused items, including the refresh button.